common template engine
