+++ cmd/infrakit/util/init/init.go
@@ -13,7 +13,6 @@ import (
 	logutil "github.com/docker/infrakit/pkg/log"
 	"github.com/docker/infrakit/pkg/plugin"
 	group_types "github.com/docker/infrakit/pkg/plugin/group/types"
-	metadata_template "github.com/docker/infrakit/pkg/plugin/metadata/template"
 	flavor_rpc "github.com/docker/infrakit/pkg/rpc/flavor"
 	metadata_rpc "github.com/docker/infrakit/pkg/rpc/metadata"
 	"github.com/docker/infrakit/pkg/run/manager"
@@ -115,7 +114,7 @@ func Command(scp scope.Scope) *cobra.Command {
 
 		if len(*metadatas) > 0 {
 			log.Info("Setting metadata entries")
-			mfunc := metadata_template.MetadataFunc(scp)
+			mfunc := scope.MetadataFunc(scp)
 			for _, md := range *metadatas {
 				// TODO -- this is not transactional.... we don't know
 				// the paths and there may be changes to multiple metadata
